requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.4.0
django-crispy-forms==2.2
crispy-bootstrap5==2024.2
django-filter==24.3
